Convert all towers from Fabric to Forge

+++ src/main/java/com/lamvdavid/pptowers/entities/TrajectoryTowerEntity.java
@@ -4,7 +4,6 @@ import net.minecraft.dispenser.Position;
 import net.minecraft.entity.projectile.ProjectileEntity;
 import net.minecraft.tileentity.TileEntityType;
 import net.minecraft.util.math.MathHelper;
-import net.minecraft.world.World;
 
 public abstract class TrajectoryTowerEntity extends ProjectileTowerEntity{
     public TrajectoryTowerEntity(TileEntityType<?> type, double xRange, double yRange, double zRange, int fireRate) {
@@ -18,7 +17,7 @@ public abstract class TrajectoryTowerEntity extends ProjectileTowerEntity{
         double x = target.getX() - proj.getX();
         double y = target.getY(0.33333333D) - proj.getY();
         double z = target.getZ() - proj.getZ();
-        double d = (double) MathHelper.sqrt(x * x + z * z);
+        double d = MathHelper.sqrt(x * x + z * z);
         proj.shoot(x, y + d * 0.20000000298023224D, z, 1.6F, 0);
         level.addFreshEntity(proj);
     }
